Paper/scripts/requirements.txt
numpy
scipy
pandas
matplotlib
jupyter
nbconvert
pyyaml
pdfplumber
google-cloud-bigquery
